+++ src/rules/unit-case/index.js
@@ -1,7 +1,7 @@
 import valueParser from "postcss-value-parser"
 import {
   declarationValueIndex,
-  isStandardValue,
+  getUnitFromValueNode,
   report,
   ruleMessages,
   validateOptions,
@@ -25,18 +25,11 @@ export default function (expectation) {
     if (!validOptions) { return }
 
     root.walkDecls(decl => {
-      const { value } = decl
-
-      valueParser(value).walk((node) => {
+      valueParser(decl.value).walk((node) => {
         // Ignore wrong units within `url` function
         if (node.type === "function" && node.value.toLowerCase() === "url") { return false }
-        if (node.type !== "word" || !isStandardValue(node.value)) { return }
 
-        const parsedUnit = valueParser.unit(node.value)
-
-        if (!parsedUnit) { return }
-
-        const unit = parsedUnit.unit
+        const unit = getUnitFromValueNode(node)
 
         if (!unit) { return }