+++ src/rules/unit-no-unknown/index.js
@@ -2,7 +2,7 @@ import { isString } from "lodash"
 import valueParser from "postcss-value-parser"
 import {
   declarationValueIndex,
-  isStandardValue,
+  getUnitFromValueNode,
   isKnownUnit,
   report,
   ruleMessages,
@@ -28,18 +28,11 @@ export default function (actual, options) {
     if (!validOptions) { return }
 
     root.walkDecls(decl => {
-      const { value } = decl
-      // Remove all interpolation inside functions
-      valueParser(value.replace(/#{.+?}|@{.+?}|\$\(.+?\)/, "")).walk(function (node) {
+      valueParser(decl.value).walk(function (node) {
         // Ignore wrong units within `url` function
         if (node.type === "function" && node.value.toLowerCase() === "url") { return false }
-        if (node.type !== "word" || !isStandardValue(node.value)) { return }
 
-        const parsedUnit = valueParser.unit(node.value)
-
-        if (!parsedUnit) { return }
-
-        const unit = parsedUnit.unit
+        const unit = getUnitFromValueNode(node)
 
         if (!unit || (unit && isKnownUnit(unit))) { return }